+++ cmake/toolchains/windows-x86_64-clang.cmake
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: LGPL-3.0-only
 
 set(CMAKE_SYSTEM_NAME
-    Linux
+    Windows
 )
 
 set(CMAKE_SYSTEM_PROCESSOR
@@ -10,23 +10,31 @@ set(CMAKE_SYSTEM_PROCESSOR
 )
 
 set(CPACK_SYSTEM_NAME
-    linux-x86_64
+    windows-x86_64
 )
 
 set(TARGET_OS
-    x86_64-pc-linux-gnu
+    x86_64-w64-mingw32
 )
 
 set(GNURX_TARGET
     ${TARGET_OS}
 )
 
+set(GNURX_CFLAGS
+    --target=${GNURX_TARGET}
+)
+
 set(FILE_TARGET
-    ${TARGET_OS}
+    x86_64-pc-windows
+)
+
+set(FILE_CFLAGS
+    --target=${TARGET_OS}
 )
 
 set(CMAKE_C_COMPILER
-    gcc
+    clang
 )
 
 set(CMAKE_C_COMPILER_TARGET
@@ -34,7 +42,7 @@ set(CMAKE_C_COMPILER_TARGET
 )
 
 set(CMAKE_CXX_COMPILER
-    g++
+    clang++
 )
 
 set(CMAKE_CXX_COMPILER_TARGET
@@ -55,6 +63,7 @@ set(CMAKE_FIND_ROOT_PATH_MODE_INCLUDE
 
 add_link_options(
     -fuse-ld=lld
+    -stdlib=libstdc++
 )
 
 add_compile_options(
@@ -62,4 +71,9 @@ add_compile_options(
     -Wextra
     -Wpedantic
     -Wfatal-errors
+    -stdlib=libstdc++
+)
+
+link_libraries(
+    pthread
 )